-renamed colour directory to brick_data -moved ldrawObject.py to brick_data dir -added main_color variable and setter to LdrawObject class -main_color is used during conversion to LDraw file
